+++ lib/zoi/types/datetime.ex
@@ -190,4 +190,8 @@ defmodule Zoi.Types.DateTime do
     defp maybe_add(map, _key, nil), do: map
     defp maybe_add(map, key, {value, _opts}), do: Map.put(map, key, DateTime.to_iso8601(value))
   end
+
+  defimpl Zoi.Describe.Encoder do
+    def encode(_schema), do: "`t:DateTime.t/0`"
+  end
 end