+++ lib/zoi/types/float.ex
@@ -195,4 +195,8 @@ defmodule Zoi.Types.Float do
     defp maybe_add(map, _key, nil), do: map
     defp maybe_add(map, key, {value, _opts}), do: Map.put(map, key, value)
   end
+
+  defimpl Zoi.Describe.Encoder do
+    def encode(_schema), do: "`t:float/0`"
+  end
 end