+++ lib/zoi/types/naive_datetime.ex
@@ -190,4 +190,8 @@ defmodule Zoi.Types.NaiveDateTime do
     defp maybe_add(map, key, {value, _opts}),
       do: Map.put(map, key, NaiveDateTime.to_iso8601(value))
   end
+
+  defimpl Zoi.Describe.Encoder do
+    def encode(_schema), do: "`t:NaiveDateTime.t/0`"
+  end
 end