+++ lib/zoi/types/tuple.ex
@@ -88,4 +88,11 @@ defmodule Zoi.Types.Tuple do
       }
     end
   end
+
+  defimpl Zoi.Describe.Encoder do
+    def encode(%{fields: fields}) do
+      types = Enum.map_join(fields, ", ", &Zoi.Describe.Encoder.encode/1)
+      "tuple of #{types} values"
+    end
+  end
 end